Move NavigationArrows from shared/components to products/components
- Fix backwards dependency: NavigationArrows in shared importing from products
- NavigationArrows only used by ProductRow, so belongs in products feature
- Update import paths: NavigationArrows self-import and ProductRow import
- Eliminates shared → products dependency violation
